refactor: 🧹 remove authentication-related logic from partnerController and userController
